Split out and simplify solve parts of IAPWS
